remove libgit2 implementation, move vcs/git/gogit -> vcs/git
Also backported a few more fallbacks.
